+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- `run_nava` in `nava.functions.py`
+- `play_cli` function
 - `SECURITY.md`
 ### Changed
 - `main` function in `__main__.py`